fix!: functions should call when pushed onto the stack (#1)
this PR corrects a misconception i had about the way a forth should work: in order to have the "left to right" concatenative semantics, functions must be called when they are pushed onto the stack.

i also felt wasteful about my use of the dictionary and pad in the previous version of hither. this PR replaces that memory model with a heap of bounded size, equipped with a mem.Allocator and implements a "mark and sweep" style garbage collector that runs when hither's state is flushed (i.e. in the provided REPL implementation, after a full line is processed and found to be complete, or on error).

the new dictionary is a hash map of bounded capacity. the keys to the hash map are strings, whose lifetimes are either tied to the initial allocation of hither (where they are copied over from static-lifetime strings in hither_lib.keys) or to the lifetime of the heap, where their presence in the hash map prevents them from being garbage-collected. the values in the hash map are single Cell objects, which are either of type builtin or of type slice with slice_is == .definition. for the latter type, the pointed-to slice is on the heap, and this indirection allows for proper variable usage semantics.

BREAKING:

, is removed
name is replaced by :=, which performs both allocation and assignment of dictionary keys. := operates both on single-word strings as 1 "x" := creating or assigning the value of 1 to the variable x or on words which are "escaped" with ' as in 1 ' x :=.
other inconsistencies between the previous collection of builtins and this collection are possible.